Split profile into sub package
