+++ src/scitacean/_profile/__init__.py
@@ -1,0 +1,7 @@
+# SPDX-License-Identifier: BSD-3-Clause
+# Copyright (c) 2026 SciCat Project (https://github.com/SciCatProject/scitacean)
+"""Profiles for connecting to SciCat."""
+
+from ._common import Profile, gather_login_params, locate_profile
+
+__all__ = ["Profile", "gather_login_params", "locate_profile"]
